Added generation code for the so-called bridge types.
These are objects that are intended to be passed from Javascript to native. Because these objects will be ultimately interfacing with statically compiled code, some amount of type checking is required.

This commit generates a bridge type module which uses a BridgeUtils object and schema to validate instances of this type.